Added first test for a PNP
The PNP is being constructed within an n-well which is deeper than the
inner p-well.
Also the outer N-well resides on a silicon island (STI isolation).

Hagen said he wants control over the potential of the base, that's what
the big half ring around the PNP is for.

Lets see what Hagen says about it.