Added first test for a NPN
The NPN is being constructed within an n-well which is deeper than the inner p-well.
Also the outer N-well resides on a silicon island (STI isolation).

Hagen said he wants control over the potential of the base, but I can't
make the big half ring like with the PNP.

Lets see what Hagen says about it.