Adding masks for our vendor
We didn't have to sign an NDA, so I think it's alright.

Also the machine for which this frame is ment is so old that it's not
bleeding edge anymore anyway.

We need to generate mask sets for the machine in the team, so the vendor
of the machines should come to me if they have a problem with me pushing
these layouts...